Feat auto fillin mf (#989)
* chart

* Update FormatCalc.java

* be: chart and style

* Fix formula evaluation and variable replacement logic

* be

* 180d > 90d

* feat: ApprovalSubmitUser

* be: AutoFillin

* be

* Update QueryParser.java

* Update charts.js

* be

* be: fixedColumns showLineNo

* be

* feat: passwd safe

* feat: execOnMainUpdate; break: GROUPAGGREGATION and mores

* be: column len

* be: execOnMainUpdate

+++ src/main/java/com/rebuild/core/service/approval/ApprovalHelper.java
@@ -283,7 +283,7 @@ public class ApprovalHelper {
         final int expiresAuto = eaConf == null ? 0 : eaConf.getIntValue("expiresAuto");
 
         if (expiresAuto == 1) {
-            int auto1Value = Math.max(eaConf.getIntValue("expiresAuto1Value"), 1);
+            int auto1Value = Math.max((int) eaConf.getDoubleValue("expiresAuto1Value"), 1);
             String auto1Type = eaConf.getString("expiresAuto1ValueType");
             // I,H,D
             return CalendarUtils.add(createdOn, auto1Value,
